refactor(copilotsdk): address manual review feedback from PR #945
- Trim copilotsdk section in providers.md to a brief one-liner referring
  to the SDK documentation, matching the style of the existing Gemini note.

- Limit auth to COPILOT_GITHUB_TOKEN and the CLI session only.
  Remove GITHUB_TOKEN and GH_TOKEN fallbacks (automatically set by GitHub
  Actions; could supply tokens that lack Copilot scopes). Add explicit
  tests confirming neither var is picked up.

- Extract module-level helpers into a new utils.py following the pattern
  used by 21 other providers. Update copilotsdk.py to import from utils.

- Fix test imports and log-warning assertions: caplog-based capture is
  unreliable when the any_llm logger has propagate=False (set by
  setup_logger() at import time in the full suite). Replace with
  patch.object on the module logger for deterministic assertion.
